requirements.txt
aiohttp
requests
wmi
python-dateutil